+++ R/RcppExports.R
@@ -61,8 +61,20 @@
     .Call(`_Cyclops_cyclopsGetLogLikelihood`, inRcppCcdInterface)
 }
 
+.cyclopsMakeDirty <- function(inRcppCcdInterface) {
+    invisible(.Call(`_Cyclops_cyclopsMakeDirty`, inRcppCcdInterface))
+}
+
 .cyclopsLogResults <- function(inRcppCcdInterface, fileName, withASE) {
     invisible(.Call(`_Cyclops_cyclopsLogResult`, inRcppCcdInterface, fileName, withASE))
+}
+
+.cyclopsTestProportionality <- function(inRcppCcdInterface, sexpBitCovariates, covariate) {
+    .Call(`_Cyclops_cyclopsTestProportionality`, inRcppCcdInterface, sexpBitCovariates, covariate)
+}
+
+.cyclopsGetSchoenfeldResiduals <- function(inRcppCcdInterface, sexpBitCovariates) {
+    .Call(`_Cyclops_cyclopsGetSchoenfeldResiduals`, inRcppCcdInterface, sexpBitCovariates)
 }
 
 .cyclopsGetFisherInformation <- function(inRcppCcdInterface, sexpBitCovariates) {
@@ -93,6 +105,14 @@
     .Call(`_Cyclops_cyclopsPredictModel`, inRcppCcdInterface)
 }
 
+.cyclopsSetConvergenceType <- function(inRcppCcdInterface, convergenceType) {
+    invisible(.Call(`_Cyclops_cyclopsSetConvergenceType`, inRcppCcdInterface, convergenceType))
+}
+
+.cyclopsGetConvergenceType <- function(inRcppCcdInterface) {
+    .Call(`_Cyclops_cyclopsGetConvergenceType`, inRcppCcdInterface)
+}
+
 .cyclopsSetControl <- function(inRcppCcdInterface, maxIterations, tolerance, convergenceType, useAutoSearch, fold, foldToCompute, lowerLimit, upperLimit, gridSteps, noiseLevel, threads, seed, resetCoefficients, startingVariance, useKKTSwindle, swindleMultipler, selectorType, initialBound, maxBoundCount, algorithm, doItAll, syncCV) {
     invisible(.Call(`_Cyclops_cyclopsSetControl`, inRcppCcdInterface, maxIterations, tolerance, convergenceType, useAutoSearch, fold, foldToCompute, lowerLimit, upperLimit, gridSteps, noiseLevel, threads, seed, resetCoefficients, startingVariance, useKKTSwindle, swindleMultipler, selectorType, initialBound, maxBoundCount, algorithm, doItAll, syncCV))
 }
@@ -115,6 +135,10 @@
 
 .cyclopsInitializeModel <- function(inModelData, modelType, computeDevice, computeMLE = FALSE) {
     .Call(`_Cyclops_cyclopsInitializeModel`, inModelData, modelType, computeDevice, computeMLE)
+}
+
+.cyclopsGetLogLikelihoodHessianDiagonal <- function(inRcppCcdInterface, sexpCovariates) {
+    .Call(`_Cyclops_cyclopsGetLogLikelihoodHessianDiagonal`, inRcppCcdInterface, sexpCovariates)
 }
 
 #' @title List available GPU devices