+++ tests/testthat/test-smallCLR.R
@@ -1,5 +1,6 @@
 library("testthat")
 library("survival")
+library("gnm")
 
 context("test-smallCLR.R")
 
@@ -42,5 +43,23 @@ test_that("Small conditional logistic regression", {
 # 	expect_equal(predict(cyclopsFit), predict(gold, type = "risk"), tolerance = tolerance)
 })
 
+test_that("Small conditional poisson regression with an offset", {
+
+    gold.cp <- gnm(event ~ exgr + agegr + offset(loginterval),
+                   family = poisson, eliminate = indiv,
+                   data = Cyclops::oxford)
+    vcov(gold.cp)
+
+    dataPtr <- createCyclopsData(event ~ exgr + agegr + strata(indiv) + offset(loginterval),
+                                 data = Cyclops::oxford,
+                                 modelType = "cpr")
+    cyclopsFit <- fitCyclopsModel(dataPtr,
+                                  prior = createPrior("none"))
+
+    expect_equal(as.numeric(coef(cyclopsFit)), as.numeric(coef(gold.cp)),
+                 tolerance = 1E-6)
+    expect_equal(as.numeric(vcov(cyclopsFit)), as.numeric(vcov(gold.cp)),
+                 tolerance = 1E-6)
+})